+++ docs/source/doxygen/html/class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.js
@@ -1,5 +1,6 @@
 var class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash =
 [
-    [ "Flash", "class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.html#a9cc6655566be1f4c88d90b2e27ada197", null ],
-    [ "IsFlashing", "class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.html#a441c884d75d143e9f4912a981cc91570", null ]
+    [ "Flash", "class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.html#a65c272b476d56a02393b5b150d794df1", null ],
+    [ "IsFlashing", "class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.html#a441c884d75d143e9f4912a981cc91570", null ],
+    [ "SetOriginalMaterial", "class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.html#a628ff2851c54302ba71a3f4e1111db40", null ]
 ];