+++ docs/source/doxygen/html/dir_395bafe2adc24fcfed5d14b8ce406a4f.js
@@ -1,4 +1,4 @@
 var dir_395bafe2adc24fcfed5d14b8ce406a4f =
 [
-    [ "UnityUtils", "dir_0a98e380e168bd97e581fa585b1ffd96.html", "dir_0a98e380e168bd97e581fa585b1ffd96" ]
+    [ "com.mydimons.unityutils", "dir_9c2fb7424bbc3a657c072801d9c5ba21.html", "dir_9c2fb7424bbc3a657c072801d9c5ba21" ]
 ];