+++ docs/source/doxygen/html/dir_3efc2c3d39c97d105de5aff9d00fb1aa.js
@@ -1,0 +1,7 @@
+var dir_3efc2c3d39c97d105de5aff9d00fb1aa =
+[
+    [ "ISaveableData.cs", "_i_saveable_data_8cs.html", "_i_saveable_data_8cs" ],
+    [ "SaveData.cs", "_save_data_8cs.html", "_save_data_8cs" ],
+    [ "SaveSlot.cs", "_save_slot_8cs.html", "_save_slot_8cs" ],
+    [ "SaveSlotSaveData.cs", "_save_slot_save_data_8cs.html", "_save_slot_save_data_8cs" ]
+];