+++ docs/source/doxygen/html/search/all_5.js
@@ -4,6 +4,6 @@ var searchData=
   ['fadeoutaudio_1',['FadeOutAudio',['../class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_animations.html#ad69193b5f38d27d69ba12fee3594bf86',1,'UnityUtils::ScriptUtils::Objects::ObjectAnimations']]],
   ['fadetime_2',['fadeTime',['../class_unity_utils_1_1_script_utils_1_1_audio_1_1_simple_background_music.html#a47f1245bae09985a4117f1af691620a7',1,'UnityUtils::ScriptUtils::Audio::SimpleBackgroundMusic']]],
   ['findalldatapersistanceobjects_3',['FindAllDataPersistanceObjects',['../class_unity_utils_1_1_script_utils_1_1_save_system_1_1_save_system_manager.html#a44ea94ab7d0fa6d624b639dc227be510',1,'UnityUtils::ScriptUtils::SaveSystem::SaveSystemManager']]],
-  ['flash_4',['Flash',['../class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.html#a9cc6655566be1f4c88d90b2e27ada197',1,'UnityUtils::ScriptUtils::Objects::ObjectColorFlash']]],
+  ['flash_4',['Flash',['../class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.html#a65c272b476d56a02393b5b150d794df1',1,'UnityUtils::ScriptUtils::Objects::ObjectColorFlash']]],
   ['flat_5',['Flat',['../namespace_unity_utils_1_1_script_utils_1_1_objects.html#a0fe45cbfcd159f24e5cedb36675c1e07a745e3db6a7ffd50e1a72b39482f0882d',1,'UnityUtils::ScriptUtils::Objects']]]
 ];