+++ docs/source/doxygen/html/search/functions_4.js
@@ -3,5 +3,5 @@ var searchData=
   ['fadeinaudio_0',['FadeInAudio',['../class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_animations.html#aff1b70dd806db24eebfb5342b29d230c',1,'UnityUtils::ScriptUtils::Objects::ObjectAnimations']]],
   ['fadeoutaudio_1',['FadeOutAudio',['../class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_animations.html#ad69193b5f38d27d69ba12fee3594bf86',1,'UnityUtils::ScriptUtils::Objects::ObjectAnimations']]],
   ['findalldatapersistanceobjects_2',['FindAllDataPersistanceObjects',['../class_unity_utils_1_1_script_utils_1_1_save_system_1_1_save_system_manager.html#a44ea94ab7d0fa6d624b639dc227be510',1,'UnityUtils::ScriptUtils::SaveSystem::SaveSystemManager']]],
-  ['flash_3',['Flash',['../class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.html#a9cc6655566be1f4c88d90b2e27ada197',1,'UnityUtils::ScriptUtils::Objects::ObjectColorFlash']]]
+  ['flash_3',['Flash',['../class_unity_utils_1_1_script_utils_1_1_objects_1_1_object_color_flash.html#a65c272b476d56a02393b5b150d794df1',1,'UnityUtils::ScriptUtils::Objects::ObjectColorFlash']]]
 ];